Refactor source files
