--- kernel/src/kernel/data_structs/partitioned_object.cpp
@@ -1,10 +1,0 @@
-#include "kernel/data_structs/partitioned_object.h"
-
- #include <stdexcept>
-
-namespace kernel
-{
-
-
-
-}
