--- kernel/src/kernel/executors/stoppable_task.cpp
@@ -1,8 +1,0 @@
-#include "kernel/executors/stoppable_task.h"
-
-namespace kernel
-{
-
-
-
-}
