--- kernel/src/kernel/kmeans/kmeans.cpp
@@ -1,11 +1,0 @@
-#include "kmeans/kmeans.h"
-
-namespace kmeans
-{
-
-	KMeans::KMeans(const KMeans::config_t& conf)
-	:
-	config_(conf)
-	{}
-
-}
