+++ kernel/src/kernel/parallel/threading/stoppable_task.cpp
@@ -1,0 +1,8 @@
+#include "kernel/parallel/threading/stoppable_task.h"
+
+namespace kernel
+{
+
+
+
+}
