+++ kernel/src/kernel/parallel/utilities/partitioned_object.cpp
@@ -1,0 +1,10 @@
+#include "kernel/parallel/utilities/partitioned_type.h"
+
+ #include <stdexcept>
+
+namespace kernel
+{
+
+
+
+}
